bigint changes

+++ services/horizon/internal/db2/schema/migrations/1_initial_schema.sql
@@ -86,7 +86,7 @@ CREATE TABLE history_ledgers (
     fee_pool bigint NOT NULL,
     base_fee integer NOT NULL,
     base_percentage_fee INTEGER,
-    max_fee INTEGER,
+    max_fee bigint,
     base_reserve integer NOT NULL,
     max_tx_set_size integer NOT NULL
 );
